[perf] Migrate @azure-tests/perf-data-tables to ESM
