+++ sdk/tables/perf-tests/data-tables/package.json
@@ -3,49 +3,45 @@
   "sdk-type": "perf-test",
   "version": "1.0.0",
   "description": "",
-  "main": "",
+  "main": "./dist/commonjs/index.js",
+  "module": "./dist/esm/index.js",
   "keywords": [],
-  "author": "",
-  "license": "ISC",
+  "author": "Microsoft Corporation",
+  "license": "MIT",
   "dependencies": {
     "@azure-tools/test-perf": "^1.0.0",
-    "@azure/core-util": "^1.6.1",
-    "@azure/data-tables": "^13.1.0",
-    "dotenv": "^16.0.0"
+    "@azure/core-util": "^1.11.0",
+    "@azure/data-tables": "^13.3.0",
+    "dotenv": "^16.0.0",
+    "tslib": "^2.8.1"
   },
   "devDependencies": {
-    "@azure-tools/test-utils-vitest": "^1.0.0",
     "@azure/dev-tool": "^1.0.0",
     "@types/node": "^18.0.0",
-    "@vitest/browser": "^2.1.8",
-    "@vitest/coverage-istanbul": "^2.1.8",
     "eslint": "^9.9.0",
-    "playwright": "^1.49.1",
-    "tslib": "^2.2.0",
-    "typescript": "~5.6.2",
-    "vitest": "^2.1.8"
+    "typescript": "~5.6.2"
   },
   "private": true,
   "scripts": {
-    "build": "npm run clean && dev-tool run build-package && dev-tool run extract-api",
+    "build": "npm run clean && dev-tool run build-package",
     "build:samples": "echo skipped",
     "build:test": "echo skipped",
-    "check-format": "dev-tool run vendored prettier --list-different --config ../../../../.prettierrc.json --ignore-path ../../../../.prettierignore \"test/**/*.ts\" \"*.{js,json}\"",
+    "check-format": "dev-tool run vendored prettier --list-different --config ../../../../.prettierrc.json --ignore-path ../../../../.prettierignore \"src/**/*.ts\" \"*.{js,json}\"",
     "clean": "dev-tool run vendored rimraf --glob dist dist-* types *.tgz *.log",
-    "format": "dev-tool run vendored prettier --write --config ../../../../.prettierrc.json --ignore-path ../../../../.prettierignore \"test/**/*.ts\" \"*.{js,json}\"",
+    "format": "dev-tool run vendored prettier --write --config ../../../../.prettierrc.json --ignore-path ../../../../.prettierignore \"src/**/*.ts\" \"*.{js,json}\"",
     "integration-test": "echo skipped",
     "integration-test:browser": "echo skipped",
     "integration-test:node": "echo skipped",
-    "lint": "eslint -c ../../../../common/tools/eslint-plugin-azure-sdk/eslint.perftests.config.mjs test",
-    "lint:fix": "eslint -c ../../../../common/tools/eslint-plugin-azure-sdk/eslint.perftests.config.mjs test --fix --fix-type [problem,suggestion]",
+    "lint": "dev-tool run vendored eslint -c ../../../../common/tools/eslint-plugin-azure-sdk/eslint.perftests.config.mjs src",
+    "lint:fix": "dev-tool run vendored eslint -c ../../../../common/tools/eslint-plugin-azure-sdk/eslint.perftests.config.mjs src --fix --fix-type [problem,suggestion]",
     "pack": "npm pack 2>&1",
-    "perf-test:node": "npm run build && node dist-esm/index.spec.js",
+    "perf-test:node": "npm run build && node dist/esm/index.js",
     "test": "echo skipped",
     "test:browser": "echo skipped",
     "test:node": "echo skipped",
     "unit-test": "echo skipped",
-    "unit-test:browser": "npm run clean && dev-tool run build-package && dev-tool run build-test && dev-tool run test:vitest --browser",
-    "unit-test:node": "dev-tool run test:vitest",
+    "unit-test:browser": "echo skipped",
+    "unit-test:node": "echo skipped",
     "update-snippets": "echo skipped"
   },
   "type": "module",
@@ -58,10 +54,6 @@
       "esm",
       "commonjs"
     ],
-    "esmDialects": [
-      "browser",
-      "react-native"
-    ],
     "selfLink": false
   },
   "files": [
@@ -69,5 +61,18 @@
     "README.md",
     "LICENSE"
   ],
-  "browser": "./dist/browser/index.js"
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "import": {
+        "types": "./dist/esm/index.d.ts",
+        "default": "./dist/esm/index.js"
+      },
+      "require": {
+        "types": "./dist/commonjs/index.d.ts",
+        "default": "./dist/commonjs/index.js"
+      }
+    }
+  },
+  "types": "./dist/commonjs/index.d.ts"
 }
