+++ sdk/tables/perf-tests/data-tables/src/listComplexEntities.spec.ts
@@ -5,14 +5,13 @@ import { PerfOptionDictionary } from "@azure-tools/test-perf";
 import { TablesTest } from "./tables.spec.js";
 import { TableEntityResult, TransactionAction } from "@azure/data-tables";
 import { createBatch } from "./utils/createBaseEntity.js";
-import { describe, it, assert } from "vitest";
 
-interface ListSimpleEntitiesTestOptions {
+interface ListComplexEntitiesTestOptions {
   entityCount: number;
 }
 
-export class ListSimpleEntitiesTest extends TablesTest<ListSimpleEntitiesTestOptions> {
-  public options: PerfOptionDictionary<ListSimpleEntitiesTestOptions> = {
+export class ListComplexEntitiesTest extends TablesTest<ListComplexEntitiesTestOptions> {
+  public options: PerfOptionDictionary<ListComplexEntitiesTestOptions> = {
     entityCount: {
       defaultValue: 100,
       longName: "entityCount",
@@ -22,13 +21,13 @@ export class ListSimpleEntitiesTest extends TablesTest<ListSimpleEntitiesTestOpt
   };
 
   constructor() {
-    super("ListSimpleEntityPerf");
+    super("ListComplexEntityPerf");
   }
 
   public async globalSetup() {
     await super.globalSetup(); // Calling base class' setup
     const batches: TransactionAction[][] = createBatch(
-      "simple",
+      "complex",
       this.parsedOptions.entityCount.value!,
     );
