+++ sdk/tables/perf-tests/data-tables/src/listSimpleEntities.spec.ts
@@ -5,14 +5,13 @@ import { PerfOptionDictionary } from "@azure-tools/test-perf";
 import { TablesTest } from "./tables.spec.js";
 import { TableEntityResult, TransactionAction } from "@azure/data-tables";
 import { createBatch } from "./utils/createBaseEntity.js";
-import { describe, it, assert } from "vitest";
 
-interface ListComplexEntitiesTestOptions {
+interface ListSimpleEntitiesTestOptions {
   entityCount: number;
 }
 
-export class ListComplexEntitiesTest extends TablesTest<ListComplexEntitiesTestOptions> {
-  public options: PerfOptionDictionary<ListComplexEntitiesTestOptions> = {
+export class ListSimpleEntitiesTest extends TablesTest<ListSimpleEntitiesTestOptions> {
+  public options: PerfOptionDictionary<ListSimpleEntitiesTestOptions> = {
     entityCount: {
       defaultValue: 100,
       longName: "entityCount",
@@ -22,13 +21,13 @@ export class ListComplexEntitiesTest extends TablesTest<ListComplexEntitiesTestO
   };
 
   constructor() {
-    super("ListComplexEntityPerf");
+    super("ListSimpleEntityPerf");
   }
 
   public async globalSetup() {
     await super.globalSetup(); // Calling base class' setup
     const batches: TransactionAction[][] = createBatch(
-      "complex",
+      "simple",
       this.parsedOptions.entityCount.value!,
     );
