--- sdk/tables/perf-tests/data-tables/tsconfig.browser.config.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./.tshy/build.json",
-  "include": ["./src/**/*.ts", "./src/**/*.mts", "./test/**/*.spec.ts", "./test/**/*.mts"],
-  "exclude": ["./test/**/node/**/*.ts"],
-  "compilerOptions": {
-    "outDir": "./dist-test/browser",
-    "rootDir": ".",
-    "skipLibCheck": true
-  }
-}
