+++ sdk/tables/perf-tests/data-tables/tsconfig.json
@@ -2,19 +2,19 @@
   "extends": "../../../../tsconfig",
   "compilerOptions": {
     "module": "NodeNext",
-    "noEmit": true,
-    "lib": ["dom", "esnext"],
     "moduleResolution": "NodeNext",
-    "rootDir": "."
+    "rootDir": ".",
+    "paths": {
+      "@azure-tests/perf-data-tables": ["./src/index"]
+    }
   },
-  "compileOnSave": true,
   "include": [
     "src/**/*.ts",
     "src/**/*.mts",
     "src/**/*.cts",
     "samples-dev/**/*.ts",
-    "test/**/*.ts",
-    "test/**/*.mts",
-    "test/**/*.cts"
+    "src/**/*.ts",
+    "src/**/*.mts",
+    "src/**/*.cts"
   ]
 }
