--- sdk/tables/perf-tests/data-tables/vitest.browser.config.ts
@@ -1,17 +1,0 @@
-
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-import { defineConfig, mergeConfig } from "vitest/config";
-import viteConfig from "../../../vitest.browser.shared.config.ts";
-
-export default mergeConfig(
-  viteConfig,
-  defineConfig({
-    test: {
-      include: [
-        "dist-test/browser/test/**/*.spec.js",
-      ],
-    },
-  }),
-);
