--- sdk/tables/perf-tests/data-tables/vitest.config.ts
@@ -1,15 +1,0 @@
-
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-import { defineConfig, mergeConfig } from "vitest/config";
-import viteConfig from "../../../vitest.shared.config.ts";
-
-export default mergeConfig(
-  viteConfig,
-  defineConfig({
-    test: {
-      include: ["test/**/*.spec.ts"],
-    },
-  }),
-);
